refactored location of SimilarRecipeDataAccessInterface (for implementation in similarRecipes use case and seeing a specific recipe)
