+++ src/main/java/data_access/DBSimilarRecipesDataAccessObject.java
@@ -1,10 +1,9 @@
 package data_access;
 
 import gateways.JavaHttpGateway;
-import okhttp3.*;
 import org.json.JSONArray;
 import org.json.JSONObject;
-import use_case.similar_recipes.SimilarRecipeDataAccessInterface;
+
 import java.util.ArrayList;
 
 public class DBSimilarRecipesDataAccessObject implements SimilarRecipeDataAccessInterface {
